third_party/WebKit/LayoutTests/bluetooth/requestDevice/chooser/fake-bluetooth-simulate-advertisement-received-test.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
// TODO(https://crbug.com/719826): This is a temporary test to try the
// FakeBluetoothChooser API as it is implemented. This test should be deleted
// after the feature is completed.
const test_desc = 'Ensure that the SimulateAdvertisementReceived API works ' +
    'correctly.';
const company_id = '224';
const data = new TextEncoder().encode('foo');
const manufacturerDataMap = {[company_id]: data};
const serviceDataMap = {[health_thermometer.uuid]: data};
const scanRecord = {
  name: 'Health Thermometer',
  uuids: ['generic_access', health_thermometer.uuid],
  txPower: 20,
  appearance: 100,
  manufacturerData: manufacturerDataMap,
  serviceData: serviceDataMap,
};
let fake_central;
let scanResult = {
  deviceAddress: '09:09:09:09:09:09',
  rssi: 100,
  scanRecord: {},
};

bluetooth_test(() => navigator.bluetooth.test.simulateCentral({
  state: 'powered-on'
})
    .then(_ => fake_central = _)
    // Test that scanRecord fields are indeed optional.
    .then(() => fake_central.simulateAdvertisementReceived(scanResult))
    .then(fake_peripheral => assert_true(
        fake_peripheral.address === '09:09:09:09:09:09'))
    // Test the scanRecord fields.
    .then(() => {
      scanResult.scanRecord = scanRecord;
      return fake_central.simulateAdvertisementReceived(scanResult);
    }), test_desc);
</script>
